TerminaController

+++ StabilimentoBalneare/src/ombrelloniani/model/Fedelta.java
@@ -4,9 +4,9 @@ public class Fedelta {
 	
 	private String nome;
 	private int giorni;
-	private double sconto;
+	private float sconto;
 
-	public Fedelta(String nome, int giorni, double sconto) {
+	public Fedelta(String nome, int giorni, float sconto) {
 		super();
 		this.nome = nome;
 		this.giorni = giorni;
@@ -29,11 +29,11 @@ public class Fedelta {
 		this.giorni = giorni;
 	}
 	
-	public double getSconto() {
+	public float getSconto() {
 		return sconto;
 	}
 	
-	public void setSconto(double sconto) {
+	public void setSconto(float sconto) {
 		this.sconto = sconto;
 	}
